Update Automapper Number to be "long" instead of "int" to handle numbers larger than 2147483647

+++ spark_auto_mapper/data_types/number.py
@@ -24,14 +24,14 @@ class AutoMapperNumberDataType(AutoMapperDataTypeBase):
             # parse the amount here
             column_spec = self.value.get_column_spec(
                 source_df=source_df, current_column=current_column
-            ).cast("int")
+            ).cast("long")
             return column_spec
         if source_df is not None and isinstance(self.value, AutoMapperDataTypeColumn) \
                 and dict(source_df.dtypes)[self.value.value] == "string":
             # parse the amount here
             column_spec = self.value.get_column_spec(
                 source_df=source_df, current_column=current_column
-            ).cast("int")
+            ).cast("long")
             return column_spec
         else:
             column_spec = self.value.get_column_spec(
